Move prepareQuery to utils
